dev-cit-3.0-sprint3-validation_5076: Initial draft of validation changes

+++ features/director-images-ws/src/main/java/com/intel/director/images/ImageActions.java
@@ -15,6 +15,11 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
+import org.apache.commons.lang.StringUtils;
+
+import com.intel.dcsg.cpg.validation.RegexPatterns;
+import com.intel.dcsg.cpg.validation.ValidationUtil;
+import com.intel.director.api.GenericResponse;
 import com.intel.director.api.ImageActionObject;
 import com.intel.director.api.ImageActionRequest;
 import com.intel.director.api.ImageActionResponse;
@@ -119,6 +124,12 @@ public class ImageActions {
 	@GET
 	public Response fetchImageAction(@PathParam("actionId") String actionId) {
 		try {
+			GenericResponse genericResponse= new GenericResponse();
+			if(!ValidationUtil.isValidWithRegex(actionId,RegexPatterns.UUID)){
+				genericResponse.error = "Action Id is empty or not in uuid format";
+				return Response.status(Response.Status.BAD_REQUEST)
+						.entity(genericResponse).build();
+			}
 			if(actionService.fetchImageAction(actionId)!=null){
 				return Response.ok(actionService.fetchImageAction(actionId))
 						.build();
@@ -181,17 +192,23 @@ public class ImageActions {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	@POST
-	public ImageActionResponse createImageAction(
+	public Response createImageAction(
 			ImageActionRequest imageActionRequest) {
 		ImageActionResponse imageActionResponse = new ImageActionResponse();
 		ImageActionObject imageActionObject;
+		String error=imageActionRequest.vaidate();
+		if(!StringUtils.isBlank(error)){
+			imageActionResponse.error=error;
+			return Response.status(Response.Status.BAD_REQUEST)
+					.entity(imageActionResponse).build();
+		}
 		try {
 			imageActionObject = actionService
 					.createImageAction(imageActionRequest);
 		} catch (DirectorException e) {
 			log.error("Error in createImageAction", e);
 			imageActionResponse.setError(e.getMessage());
-			return imageActionResponse;
+			return Response.ok(imageActionResponse).build();
 		}
 		imageActionResponse.setId(imageActionObject.getId());
 		imageActionResponse.setImage_id(imageActionObject.getImage_id());
@@ -204,7 +221,7 @@ public class ImageActions {
 				.getCurrent_task_name());
 		imageActionResponse.setCurrent_task_status(imageActionObject
 				.getCurrent_task_status());
-		return imageActionResponse;
+		return Response.ok(imageActionResponse).build();
 	}
 
 	/**
@@ -226,19 +243,25 @@ public class ImageActions {
 	@Path("image-actions/{actionId: [0-9a-zA-Z_-]+}")
 	@Produces(MediaType.APPLICATION_JSON)
 	@DELETE
-	public ImageActionResponse deleteImageAction(
+	public Response deleteImageAction(
 			@PathParam("actionId") String actionId) {
 		ImageActionResponse imageActionResponse = new ImageActionResponse();
 		imageActionResponse.setDeleted(true);
+	///	GenericResponse genericResponse= new GenericResponse();
+		if(!ValidationUtil.isValidWithRegex(actionId,RegexPatterns.UUID)){
+			imageActionResponse.error = "Action Id is empty or not in uuid format";
+			return Response.status(Response.Status.BAD_REQUEST)
+					.entity(imageActionResponse).build();
+		}
 		try {
 			actionService.deleteImageAction(actionId);
 		} catch (Exception e) {
 			log.error("Error in deleteImageAction", e);
 			imageActionResponse.setError("Error in deleteImageAction");
 			imageActionResponse.setDeleted(false);
-			return imageActionResponse;
+			return Response.ok(imageActionResponse).build();
 		}
-		return imageActionResponse;
+		return Response.ok(imageActionResponse).build();
 	}
 
 }
