+++ features/director-images-ws/src/main/java/com/intel/director/images/TrustPolicyDrafts.java
@@ -12,15 +12,17 @@ import javax.ws.rs.core.Response;
 
 import org.apache.commons.lang.StringUtils;
 
+import com.intel.dcsg.cpg.validation.RegexPatterns;
+import com.intel.dcsg.cpg.validation.ValidationUtil;
 import com.intel.director.api.CreateTrustPolicyMetaDataRequest;
 import com.intel.director.api.CreateTrustPolicyMetaDataResponse;
 import com.intel.director.api.CreateTrustPolicyResponse;
 import com.intel.director.api.GenericRequest;
 import com.intel.director.api.GenericResponse;
 import com.intel.director.api.ListTrustPolicyDrafts;
-import com.intel.director.api.TrustPolicyDraft;
 import com.intel.director.api.TrustPolicyDraftEditRequest;
 import com.intel.director.api.ui.ImageInfo;
+import com.intel.director.api.ui.TrustPolicyDraftResponse;
 import com.intel.director.common.Constants;
 import com.intel.director.images.exception.DirectorException;
 import com.intel.director.service.ImageService;
@@ -151,8 +153,12 @@ public class TrustPolicyDrafts {
 	public Response getPolicyDraft(
 			@PathParam("trustPolicyDraftId") String trustPolicyDraftId) {
 
-		CreateTrustPolicyMetaDataResponse createTrustPolicyMetaDataResponse = null;
-
+		CreateTrustPolicyMetaDataResponse createTrustPolicyMetaDataResponse = new CreateTrustPolicyMetaDataResponse();
+		if(!ValidationUtil.isValidWithRegex(trustPolicyDraftId,RegexPatterns.UUID)){
+			createTrustPolicyMetaDataResponse.error = "Trust Policy Draft Id is empty or not in uuid format";
+			return Response.status(Response.Status.BAD_REQUEST)
+					.entity(createTrustPolicyMetaDataResponse).build();
+		}
 		try {
 			createTrustPolicyMetaDataResponse = imageService
 					.getPolicyMetadata(trustPolicyDraftId);
@@ -239,12 +245,20 @@ public class TrustPolicyDrafts {
 	@POST
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
-	public TrustPolicyDraft editPolicyDraft(
+	public Response editPolicyDraft(
 			@PathParam("trustPolicyDraftId") String trustPolicyDraftId,
 			TrustPolicyDraftEditRequest trustPolicyDraftEditRequest) {
 		trustPolicyDraftEditRequest.trust_policy_draft_id = trustPolicyDraftId;
-		TrustPolicyDraft policyDraft = null;
+		TrustPolicyDraftResponse policyDraft = new TrustPolicyDraftResponse();
 		String trustPolicyDraftXML = null;
+		String error=trustPolicyDraftEditRequest.validate();
+		if(!StringUtils.isBlank(error)){
+			policyDraft.error = error;
+			return Response.status(Response.Status.BAD_REQUEST)
+					.entity(policyDraft).build();
+		}
+	
+		
 		try {
 			policyDraft = imageService
 					.editTrustPolicyDraft(trustPolicyDraftEditRequest);
@@ -258,13 +272,13 @@ public class TrustPolicyDrafts {
 					+ trustPolicyDraftXML);
 		} catch (Exception e) {
 			// /response.setStatus(Constants.ERROR);
-			policyDraft = new TrustPolicyDraft();
 			policyDraft.setError(e.getMessage());
 			log.error("Error while updating policy draft : "
 					+ trustPolicyDraftId);
 		}
 
-		return policyDraft;
+		return 
+				Response.ok(policyDraft).build();
 	}
 
 	/**
@@ -297,24 +311,30 @@ public class TrustPolicyDrafts {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	@POST
-	public CreateTrustPolicyResponse createTrustPolicy(
+	public Response createTrustPolicy(
 			CreateTrustPolicyMetaDataRequest createPolicyRequest) {
-		CreateTrustPolicyResponse response = new CreateTrustPolicyResponse();
+		CreateTrustPolicyResponse createTrustPolicyDraftResponse = new CreateTrustPolicyResponse();
 		try {
 			ImageInfo imageInfo = imageService.fetchImageById(createPolicyRequest.image_id);
+			String error=createPolicyRequest.validate("policy");
+			if(!StringUtils.isBlank(error)){
+				createTrustPolicyDraftResponse.error = error;
+				return Response.status(Response.Status.BAD_REQUEST)
+						.entity(createTrustPolicyDraftResponse).build();
+			}
 			String imageId = imageInfo.id;
 			String trustPolicyId = imageService
 					.createTrustPolicy(imageId, createPolicyRequest.trust_policy_draft_id);
-			response.setId(trustPolicyId);
+			createTrustPolicyDraftResponse.setId(trustPolicyId);
 			imageService.deletePasswordForHost(imageId);
 			// / response.setStatus(Constants.SUCCESS);
 		} catch (DirectorException de) {
 			// //response.setStatus(Constants.ERROR);
-			response.setError(de.getMessage());
+			createTrustPolicyDraftResponse.setError(de.getMessage());
 			log.error("Error creating policy from draft for image : "
 					+ createPolicyRequest.image_id, de);
 		}
-		return response;
+		return Response.ok(createTrustPolicyDraftResponse).build();
 	}
 
 	/**
@@ -341,31 +361,30 @@ public class TrustPolicyDrafts {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	@POST
-	public CreateTrustPolicyMetaDataResponse createTrustPolicyDraft(
+	public Response createTrustPolicyDraft(
 			CreateTrustPolicyMetaDataRequest createTrustPolicyMetaDataRequest) {
 
 		CreateTrustPolicyMetaDataResponse createTrustPolicyMetadataResponse = new CreateTrustPolicyMetaDataResponse();
-		if (StringUtils.isBlank(createTrustPolicyMetaDataRequest
-				.getDisplay_name())) {
-			createTrustPolicyMetadataResponse.setError("Display Name is mandatory");
-		} else if (StringUtils.isBlank(createTrustPolicyMetaDataRequest.getLaunch_control_policy())) {
-			createTrustPolicyMetadataResponse.setError("Launch_control_policy is mandatory");
-		} else if (StringUtils.isBlank(createTrustPolicyMetaDataRequest.getImage_id())) {
-			createTrustPolicyMetadataResponse.setError("image_id is mandatory");
-		} else {
-			try {
-				createTrustPolicyMetadataResponse = imageService
-						.saveTrustPolicyMetaData(createTrustPolicyMetaDataRequest);
-			} catch (DirectorException e) {
-				log.error("createTrustPolicyMetaData failed", e);
-				// createTrustPolicyMetadataResponse.setStatus(Constants.ERROR);
-				// createTrustPolicyMetadataResponse.setDetails(e.getMessage());
-				createTrustPolicyMetadataResponse.setError("Unable to create policy draft");
-				return createTrustPolicyMetadataResponse;
-			}
+		String error = createTrustPolicyMetaDataRequest.validate("draft");
+		if (!StringUtils.isBlank(error)) {
+			createTrustPolicyMetadataResponse.setError(error);
+			return Response.status(Response.Status.BAD_REQUEST)
+			.entity(createTrustPolicyMetadataResponse).build();
 		}
 
-		return createTrustPolicyMetadataResponse;
+		try {
+			createTrustPolicyMetadataResponse = imageService
+					.saveTrustPolicyMetaData(createTrustPolicyMetaDataRequest);
+		} catch (DirectorException e) {
+			log.error("createTrustPolicyMetaData failed", e);
+			// createTrustPolicyMetadataResponse.setStatus(Constants.ERROR);
+			// createTrustPolicyMetadataResponse.setDetails(e.getMessage());
+			createTrustPolicyMetadataResponse
+					.setError("Unable to create policy draft");
+			return Response.ok(createTrustPolicyMetadataResponse).build();
+		}
+
+		return Response.ok(createTrustPolicyMetadataResponse).build();
 	}
 
 	/**
@@ -394,9 +413,14 @@ public class TrustPolicyDrafts {
 	@Path("rpc/create-draft-from-policy")
 	@POST
 	@Produces(MediaType.APPLICATION_JSON)
-	public TrustPolicyDraft createPolicyDraftFromPolicy(GenericRequest req)
+	public Response createPolicyDraftFromPolicy(GenericRequest req)
 			 {
-		TrustPolicyDraft trustPolicyDraft= new TrustPolicyDraft();
+		TrustPolicyDraftResponse trustPolicyDraft= new TrustPolicyDraftResponse();
+		if(!ValidationUtil.isValidWithRegex(req.getImage_id(),RegexPatterns.UUID)){
+			trustPolicyDraft.error = "Image id is empty or not in uuid format";
+			return Response.status(Response.Status.BAD_REQUEST)
+					.entity(trustPolicyDraft).build();
+		}
 		try {
 			
 			trustPolicyDraft= imageService.createPolicyDraftFromPolicy(req.getImage_id());
@@ -405,7 +429,7 @@ public class TrustPolicyDrafts {
 			trustPolicyDraft.setError(e.getMessage());
 			
 		}
-		return trustPolicyDraft;
+		return Response.ok(trustPolicyDraft).build();
 	}
 
 	/**
@@ -428,18 +452,23 @@ public class TrustPolicyDrafts {
 	@Path("trust-policy-drafts/{trustPolicyDraftId: [0-9a-zA-Z_-]+}")
 	@DELETE
 	@Produces(MediaType.APPLICATION_JSON)
-	public GenericResponse deletePolicyDraft(
+	public Response deletePolicyDraft(
 			@PathParam("trustPolicyDraftId") String trustPolicyDraftId) {
-		GenericResponse response = new GenericResponse();
+		GenericResponse genericResponse = new GenericResponse();
+		if(!ValidationUtil.isValidWithRegex(trustPolicyDraftId,RegexPatterns.UUID)){
+			genericResponse.error = "Trust Policy Draft Id is empty or not in uuid format";
+			return Response.status(Response.Status.BAD_REQUEST)
+					.entity(genericResponse).build();
+		}
 		try {
 			imageService.deleteTrustPolicyDraft(trustPolicyDraftId);
-			response.setDeleted(true);
+			genericResponse.setDeleted(true);
 		} catch (DirectorException e) {
 			log.error("Error in deletePolicyDraft", e);
-			response.setDeleted(false);
-			response.setError("Error in deletePolicyDraft");
+			genericResponse.setDeleted(false);
+			genericResponse.setError("Error in deletePolicyDraft");
 		}
-		return response;
+		return Response.ok(genericResponse).build();
 
 	}
 }
